React_FullStack_App/backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pymongo==4.10.1
python-jose[cryptography]==3.4.0
passlib==1.7.4
python-dotenv==1.0.1
